[mlir] Move WalkResult to Support
This also enables moving StateStack, both are relatively generic helper
structs not tied to IR.
