+++ mlir/include/mlir/IR/Visitors.h
@@ -14,6 +14,7 @@
 #define MLIR_IR_VISITORS_H
 
 #include "mlir/Support/LLVM.h"
+#include "mlir/Support/WalkResult.h"
 #include "llvm/ADT/STLExtras.h"
 
 namespace mlir {
@@ -22,41 +23,6 @@ class InFlightDiagnostic;
 class Operation;
 class Block;
 class Region;
-
-/// A utility result that is used to signal how to proceed with an ongoing walk:
-///   * Interrupt: the walk will be interrupted and no more operations, regions
-///   or blocks will be visited.
-///   * Advance: the walk will continue.
-///   * Skip: the walk of the current operation, region or block and their
-///   nested elements that haven't been visited already will be skipped and will
-///   continue with the next operation, region or block.
-class WalkResult {
-  enum ResultEnum { Interrupt, Advance, Skip } result;
-
-public:
-  WalkResult(ResultEnum result = Advance) : result(result) {}
-
-  /// Allow LogicalResult to interrupt the walk on failure.
-  WalkResult(LogicalResult result)
-      : result(failed(result) ? Interrupt : Advance) {}
-
-  /// Allow diagnostics to interrupt the walk.
-  WalkResult(Diagnostic &&) : result(Interrupt) {}
-  WalkResult(InFlightDiagnostic &&) : result(Interrupt) {}
-
-  bool operator==(const WalkResult &rhs) const { return result == rhs.result; }
-  bool operator!=(const WalkResult &rhs) const { return result != rhs.result; }
-
-  static WalkResult interrupt() { return {Interrupt}; }
-  static WalkResult advance() { return {Advance}; }
-  static WalkResult skip() { return {Skip}; }
-
-  /// Returns true if the walk was interrupted.
-  bool wasInterrupted() const { return result == Interrupt; }
-
-  /// Returns true if the walk was skipped.
-  bool wasSkipped() const { return result == Skip; }
-};
 
 /// Traversal order for region, block and operation walk utilities.
 enum class WalkOrder { PreOrder, PostOrder };
